+++ cbir/CHANGELOG.md
@@ -10,6 +10,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add H-Optimus-0 model
+- Add ResNet model by default
+
+### Removed
+
+- Remove DVC support
 
 ## [0.4.0] - 2025-05-14